Issue-184 - added further refactoring
