+++ client/build.gradle.kts
@@ -128,6 +128,7 @@ dependencies {
     implementation(libs.kotlin.coroutines.android)
     implementation(libs.kotlin.multik.core)
     debugImplementation(libs.kotlin.multik.kotlin)
+    testImplementation(libs.kotlin.multik.kotlin)
     implementation(project(":commons"))
 
     debugImplementation(libs.androidx.compose.ui.test.manifest)
