+++ client/src/testFullDebug/java/com/github/goldy1992/mp3player/client/utils/visualizer/circle/CircularCurveFitterTestBase.kt
@@ -1,0 +1,9 @@
+package com.github.goldy1992.mp3player.client.utils.visualizer.circle
+
+open class CircularCurveFitterTestBase {
+
+    open fun createFrequenciesTestCase1() : List<Float> {
+        return (1..15).map { 200f }.toList()
+    }
+
+}
